Create Workload V2: Move extractors to single module and add SynchronousCorpusExtractor
